# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AWS Image Service API
  description: |
    Serverless image upload and AI analysis service built on AWS.

    ## Authentication
    Most endpoints require JWT authentication via AWS Cognito. Include the access token
    in the Authorization header as a Bearer token.

    ## User Roles
    - **user**: Can only access their own images and analysis results
    - **admin**: Can access all users' data, with optional `?userId=` filter
  version: 1.0.0
  contact:
    name: API Support
    email: john.gambrell@gmail.com
servers:
  - url: https://{api-id}.execute-api.{region}.amazonaws.com/prod
    description: Production API Gateway
    variables:
      api-id:
        default: xxxxxxxxxx
        description: API Gateway ID
      region:
        default: us-east-1
        description: AWS Region
...
components:
  securitySchemes:
    CognitoAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT access token
  schemas:
...
    SignUpRequest:
      type: object
      required:
        - email
        - password
        - givenName
        - familyName
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: MySecureP@ss123
        givenName:
          type: string
          example: John
        familyName:
          type: string
          example: Doe
    SignInRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: MySecureP@ss123
    ConfirmSignUpRequest:
      type: object
      required:
        - email
        - confirmationCode
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        confirmationCode:
          type: string
          example: "123456"
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: eyJjdHkiOiJKV1QiLCJlbmMiOiJBMjU2R0NNIiwiYWxnIjoiUlNBLU9BRVAifQ...
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: user@example.com
    ConfirmForgotPasswordRequest:
      type: object
      required:
        - email
        - confirmationCode
        - newPassword
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        confirmationCode:
          type: string
          example: "123456"
        newPassword:
          type: string
          format: password
          minLength: 8
          example: MyNewSecureP@ss456
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token for API authorization
        idToken:
          type: string
          description: JWT ID token with user claims
        refreshToken:
          type: string
          description: Token to refresh access token (only on sign-in)
        expiresIn:
          type: integer
          description: Token expiration in seconds
          example: 3600
    User:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: Cognito user sub (unique identifier)
        email:
          type: string
          format: email
        givenName:
          type: string
        familyName:
          type: string
        role:
          type: string
          enum: [user, admin]

    # Image Schemas
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        error:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
paths:
...
  /api/auth/signup:
    post:
      tags:
        - Auth
      summary: Register new user
      description: Create a new user account. A confirmation code will be sent to the email.
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          userId:
                            type: string
                          confirmed:
                            type: boolean
                          message:
                            type: string
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/confirm:
    post:
      tags:
        - Auth
      summary: Confirm email registration
      description: Confirm user registration with the code sent to email
      operationId: confirmSignUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmSignUpRequest"
      responses:
        "200":
          description: Email confirmed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Invalid or expired confirmation code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/signin:
    post:
      tags:
        - Auth
      summary: Sign in user
      description: Authenticate user and return JWT tokens
      operationId: signIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthTokens"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Email not confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: Get new access and ID tokens using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthTokens"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Initiate password reset
      description: Send password reset code to user's email
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Password reset code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/confirm-forgot-password:
    post:
      tags:
        - Auth
      summary: Complete password reset
      description: Reset password using confirmation code
      operationId: confirmForgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmForgotPasswordRequest"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Invalid or expired confirmation code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/me:
    get:
      tags:
        - Auth
      summary: Get current user profile
      description: Get the authenticated user's profile information
      operationId: getMe
      security:
        - CognitoAuth: []
      responses:
        "200":
          description: User profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Image Endpoints
...